+++ app/src/main/java/dev/hossain/devicecatalog/ui/devicedetails/DeviceDetailsScreen.kt
@@ -21,5 +21,7 @@ data class DeviceDetailsScreen(
         data object BackClicked : Event()
 
         data object RetryLoading : Event()
+
+        data object ShareClicked : Event()
     }
 }